Production check: 42 adversarially verified fixes applied across 31 files. All 615 tests pass.
Build (9 fixes):
• Debug|x64 MKL library: added mkl_intel_lp64.lib to linker
• CMake Windows variogram: fixed missing .dll extension (SUFFIX)
• CMake Unix variogram: fixed double-underscore prefix
• CMake MKL guard: HPGL_USE_MKL compile definition after target creation
• CMake empty HPGL_BUILD_PYTHON block removed
• MSBuild: lapack_compat.h added to ClInclude
• build.bat: improved VS path detection, CLI parsing, dead endlocal removed

C++ (10 fixes):
• Dangling pointer: hpgl_get_last_exception_message returns safe copy
• Bounds check: non_parametric_cdf operator() guards empty CDF
• Rule of Three: clusterizer_t copy ops deleted
• Memory safety: unique_ptr for PIMPL, null checks, calloc checks
• Type safety: int truncation fixes, enum class modernization, const correctness
• MKL: conditional header includes (lapack_compat.h)
• Dead code: sgs_params.h operator<< unreachable return removed

Python (20+ fixes):
• Tabs-to-spaces conversion across 5 files (1270 lines)
• Struct field renames: data→m_data, strides→m_strides (consistency with C API)
• Path validation: PathValidator added to SaveGSLIBPointSet/SaveGSLIBCubes
• NumPy 2.x: row_stack→vstack, numpy>=2.0 in pyproject.toml
• Dead code: scipy star imports removed (zero scipy uses)
• DLL safety: cvariogram.py loads via safe loader
• Error handling: _snapshot_hpgl_error in simple_kriging_weights
• __all__ completeness: added 3 missing functions
• Naming: cont_point_set_t values_count→size

Tests (5 fixes):
• Improved assertions in basic test files
• Value checks in parametrized kriging tests
• MovingAverage3D value correctness assertions
• Fixed test regression from tabs conversion (assertions restored inside loop)
• Updated cokriging docstring for parameter order

Config:
• pyproject.toml: numpy >=2.0, upper bound <3.0, ruff per-file-ignores
• .gitignore: added .benchmarks/

Pipeline: 65+ agents across 5 stages (discovery → adversarial → fixes → review → verification)
615 tests pass, 0 failures, 0 skipped.

+++ src/geo_bsd/hpgl/covariance_field.cpp
@@ -91,7 +91,7 @@ void covariance_field_t::init(
 
 	double treshold = value(0, 0, 0) / 100;
 
-	for (int idx = 0, end_idx = m_vectors.size(); idx < end_idx; ++idx)
+	for (size_t idx = 0, end_idx = m_vectors.size(); idx < end_idx; ++idx)
 	{
 		sugarbox_vector_t vec = m_vectors[idx];
 		if (value(vec[0],vec[1],vec[2]) < treshold)
